oshmpi_impl: fix comparison of negative required_ext_nelems
We pass -1 for the parameter required_ext_nelems to OSHMPI_create_strided_dtype
in order to skip the MPI type creation. But checking if the unsigned type
required_ext_nelems > 0 will always result in true. Thus it is better to just
make the parameter signed type.

+++ src/include/oshmpi_impl.h
@@ -419,7 +419,7 @@ OSHMPI_STATIC_INLINE_PREFIX void OSHMPI_translate_disp_to_vaddr(OSHMPI_symm_obj_
  * The caller must check the returned datatype to free it when necessary. */
 OSHMPI_STATIC_INLINE_PREFIX void OSHMPI_create_strided_dtype(size_t nelems, ptrdiff_t stride,
                                                              MPI_Datatype mpi_type,
-                                                             size_t required_ext_nelems,
+                                                             int required_ext_nelems,
                                                              size_t * strided_cnt,
                                                              MPI_Datatype * strided_type)
 {
